+++ bindings/wisdom.h
@@ -892,6 +892,7 @@ enum WisMemoryType {
      * Used for data that is directly uploaded to the GPU Local memory using copy operations.
      * Can be used only with Extended Allocation extension enabled and supported.
      * Usage outside of the Extended Allocation is undefined behavior.
+     * Usage with buffers requires allocating memory on GPUUpload heap and mapping the memory.
      * */
     MemoryTypeGPUUpload = 3,
 };
@@ -1203,7 +1204,8 @@ enum WisMemoryFlagsBits {
     /**
      * @brief Memory is exportable.
      * If set, memory can be exported to other processes or APIs.
-     * NOTE: Can't be used with MemoryTypeGPUUpload on NVidia as of 9/7/2024. Bug was reported.
+     * Works only with Device Local memory (MemoryTypeDefault) and only on AllocateXMemory calls.
+     * Outside of AllocateXMemory the flag is ignored.
      * */
     MemoryFlagsExportable = 1 << 2,
 };